Add files via upload
Finalized code for data analysis. This code has capability to process csv files based on field, display, and date filters, or to process based on TCP and UDP conversations. It also yields capability to perform various calculations on this processed data to obtain many data points, which are primarily based on Professor Zheleva's research article (which includes Tables 1, 2, 3, 4, and 6 as well as Figures 2, 3, 4, 5, 6, 7, 8, and 9) although there are more data points that can be calculated outside of this scope. 

Finalized code date: 4/4/2022
Written and committed by: Munthir Chater